# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "BPMT API's",
        "description": "Project management operations",
        "version": "1.0.0"
    },
    "host": "http://my-default-host.com",
    "basePath": "/api/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/v1/designation": {
            "get": {
                "tags": [
                    "Designation"
                ],
                "summary": "Designation list",
                "description": "Returns designation list",
                "operationId": "list of designation",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Designation"
                ],
                "summary": "Create or update designation",
                "description": "Create or update designation",
                "operationId": "create designation",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "designation Id at the time of update",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "designation title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "company_id",
                        "in": "formData",
                        "description": "Id of the company",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/designation/{id}": {
            "get": {
                "tags": [
                    "Designation"
                ],
                "summary": "designation details",
                "description": "Returns company details",
                "operationId": "single designation",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "designation Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Designation"
                ],
                "summary": "Delete a designation",
                "description": "Delete a designation",
                "operationId": "delete designation",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Designation Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}